Move deletee.js
